Change uint_8 to byte (typedef unsiged chard), uint16_t to unsigned short, uint32_t to unsigned int, int32_t to int. Add the process_instruction function of assemble inside main and left the function pointer outside of main. Change the dependencies of emulate and assemble. Repleace the arguments of functions to pointers and constants where possible for asemble related functions.

+++ src/emulate_utils.c
@@ -1,13 +1,13 @@
 #include "emulate_utils.h"
 
 // Checks if the sum of 2 integers gives an overflow (the sum is given as an argument as well)
-uint8_t check_overflow(uint32_t const *a, uint32_t const *b, uint32_t const *result) {
-    return ((*a > 0 && *b > 0 && ((int32_t) *result) < 0) || (((int32_t) *a) < 0 && ((int32_t) *b) < 0 && *result > 0));
+byte check_overflow(unsigned int const *a, unsigned int const *b, unsigned int const *result) {
+    return ((*a > 0 && *b > 0 && ((int) *result) < 0) || (((int) *a) < 0 && ((int) *b) < 0 && *result > 0));
 }
 
 // Check if Cond satisfies CPSR register [Pg. 4 spec]
-uint8_t is_condition_satisfied(uint32_t const *cpsr, uint8_t const *condition) {
-    uint8_t important_bits = *cpsr >> 28;
+byte is_condition_satisfied(unsigned int const *cpsr, byte const *condition) {
+    byte important_bits = *cpsr >> 28;
     switch (*condition) {
         case EQ: // (Z == 1)
             return important_bits & 0x4;
@@ -28,8 +28,8 @@ uint8_t is_condition_satisfied(uint32_t const *cpsr, uint8_t const *condition) {
     }
 }
 
-void shift_with_carry(uint8_t const *shift_by, uint8_t const *shift_type,
-                      uint32_t *shifted_value, uint8_t *shift_carry) {
+void shift_with_carry(byte const *shift_by, byte const *shift_type,
+                      unsigned int *shifted_value, byte *shift_carry) {
     switch (*shift_type) {
         case LSL:
             *shift_carry = (*shifted_value >> (32 - *shift_by)) & 0x1;
@@ -41,7 +41,7 @@ void shift_with_carry(uint8_t const *shift_by, uint8_t const *shift_type,
             break;
         case ASR:
             *shift_carry = (*shifted_value >> (*shift_by - 1)) & 0x1;
-            *shifted_value = (uint32_t) (((int32_t) *shifted_value) >> *shift_by);
+            *shifted_value = (unsigned int) (((int) *shifted_value) >> *shift_by);
             break;
         case ROR:
             *shift_carry = (*shifted_value >> (*shift_by - 1)) & 0x1;
@@ -52,7 +52,7 @@ void shift_with_carry(uint8_t const *shift_by, uint8_t const *shift_type,
     }
 }
 
-void shift(uint8_t const *shift_by, uint8_t const *shift_type, uint32_t *shifted_value) {
+void shift(byte const *shift_by, byte const *shift_type, unsigned int *shifted_value) {
     switch (*shift_type) {
         case LSL:
             *shifted_value = *shifted_value << *shift_by;
@@ -61,7 +61,7 @@ void shift(uint8_t const *shift_by, uint8_t const *shift_type, uint32_t *shifted
             *shifted_value = *shifted_value >> *shift_by;
             break;
         case ASR:
-            *shifted_value = (uint32_t) (((int32_t) *shifted_value) >> *shift_by);
+            *shifted_value = (unsigned int) (((int) *shifted_value) >> *shift_by);
             break;
         case ROR:
             *shifted_value = (*shifted_value >> *shift_by) | (*shifted_value << (32 - *shift_by));
@@ -71,17 +71,17 @@ void shift(uint8_t const *shift_by, uint8_t const *shift_type, uint32_t *shifted
     }
 }
 
-void processing_update_CPSR(uint32_t *cpsr, uint32_t const *result, uint8_t const *carry) {
+void processing_update_CPSR(unsigned int *cpsr, unsigned int const *result, byte const *carry) {
     if (*result) {
         if (*carry) {
-            uint8_t bit31 = (*result >> 31) & 0x1;
+            byte bit31 = (*result >> 31) & 0x1;
             if (bit31) {
                 *cpsr = (*cpsr & 0xbfffffff) | 0xa0000000; // clears Z and sets N and C of CPSR
             } else {
                 *cpsr = (*cpsr & 0x3fffffff) | 0x20000000; // clears N and Z and sets C of CPSR
             }
         } else {
-            uint8_t bit31 = (*result >> 31) & 0x1;
+            byte bit31 = (*result >> 31) & 0x1;
             if (bit31) {
                 *cpsr = (*cpsr & 0x9fffffff) | 0x80000000; // clears Z and C and sets N of CPSR
             } else {
@@ -97,9 +97,9 @@ void processing_update_CPSR(uint32_t *cpsr, uint32_t const *result, uint8_t cons
     }
 }
 
-void multiplying_update_CPSR(uint32_t *cpsr, uint32_t const *result) {
+void multiplying_update_CPSR(unsigned int *cpsr, unsigned int const *result) {
     if (result) {
-        uint8_t bit31 = (*result >> 31) & 0x1;
+        byte bit31 = (*result >> 31) & 0x1;
         if (bit31) {
             *cpsr = (*cpsr & 0x8fffffff) | 0x80000000; // clears Z and sets N of CPSR
         } else {
@@ -110,7 +110,7 @@ void multiplying_update_CPSR(uint32_t *cpsr, uint32_t const *result) {
     }
 }
 
-void gpio_access_print(uint32_t *address) {
+void gpio_access_print(unsigned int *address) {
     if (*address <= GPIO_0_9_LAST_BYTE) {
         printf("One GPIO pin from 0 to 9 has been accessed\n");
     } else if (*address <= GPIO_10_19_LAST_BYTE) {
@@ -120,24 +120,24 @@ void gpio_access_print(uint32_t *address) {
     }
 }
 
-void load(uint32_t const *address, uint8_t const *memory, uint32_t *Rd_register) {
+void load(unsigned int const *address, byte const *memory, unsigned int *Rd_register) {
     *Rd_register = (((((memory[*address + 3] << 8) |
                        memory[*address + 2]) << 8) |
                      memory[*address + 1]) << 8) |
                    memory[*address];
 }
 
-void store(uint32_t const *address, uint8_t *memory, uint32_t const *Rd_register) {
+void store(unsigned int const *address, byte *memory, unsigned int const *Rd_register) {
     memory[*address] = *Rd_register & 0xff;
     memory[*address + 1] = (*Rd_register >> 8) & 0xff;
     memory[*address + 2] = (*Rd_register >> 16) & 0xff;
     memory[*address + 3] = (*Rd_register >> 24) & 0xff;
 }
 
-void output(uint32_t const *registers, uint8_t const *memory) {
+void output(unsigned int const *registers, byte const *memory) {
     // Print registers
     printf("Registers:\n");
-    for (int i = 0; i < 13; i++) {
+    for (byte i = 0; i < 13; i++) {
         printf("$%-3d: %10d (0x%08x)\n", i, registers[i], registers[i]);
     }
     printf("PC  : %10d (0x%08x)\n", registers[PC_REF], registers[PC_REF]);
@@ -145,11 +145,11 @@ void output(uint32_t const *registers, uint8_t const *memory) {
 
     // Print non-zero memory
     printf("Non-zero memory:\n");
-    for (uint32_t i = 0; i < MEMORY_SIZE; i += 4) {
-        int32_t mem = (((((memory[i] << 8) |
-                          memory[i + 1]) << 8) |
-                        memory[i + 2]) << 8) |
-                      memory[i + 3];
+    for (unsigned int i = 0; i < MEMORY_SIZE; i += 4) {
+        int mem = (((((memory[i] << 8) |
+                      memory[i + 1]) << 8) |
+                    memory[i + 2]) << 8) |
+                  memory[i + 3];
         if (mem != 0) {
             printf("0x%08x: 0x%08x\n", i, mem);
         }
